[feature] Parallelize list-roles (#47)
Listing roles and alias across a bunch of accounts was getting slow and making our webserver take a long time to boot. Parallelized that work. 

Note that I added a mutex to protect the map we're aggregating to. At some point we should refactor this logic to clean that up, but feels ok to keep as-is for now.